Agregando mensajes flash, confirmaciónes y Docker
Se agrego el uso de docker y docker compose para ejecutar el proyecto

+++ front-agenda/src/components/DisponibilidadUsuario.vue
@@ -65,8 +65,10 @@ export default {
     },
     methods: {
     eliminarDisponibilidad(id) {
-        const store = useStore();
-        store.deleteDisponibilidad(id);
+        if (confirm('¿Estás seguro de que quieres eliminar esta disponibilidad?')) {
+            const store = useStore();
+            store.deleteDisponibilidad(id);
+        }
     },
     async verInfo(disponibilidad) {
         try {
